+++ src/commands/__init__.py
@@ -1,15 +1,17 @@
 from ...override_audit import reload
 
 reload("src.commands", ["package_report", "override_report", "diff_report",
-       "refresh_report", "diff_single", "toggle_override", "diff_override",
-       "edit_override", "delete_override", "freshen_override", "diff_package",
-       "diff_externally", "revert_override", "freshen_package", "modify_mark"])
+       "refresh_report", "diff_single", "toggle_override", "create_override",
+       "diff_override", "edit_override", "delete_override", "freshen_override",
+       "diff_package", "diff_externally", "revert_override", "freshen_package",
+       "modify_mark"])
 
 from .package_report import OverrideAuditPackageReportCommand
 from .override_report import OverrideAuditOverrideReportCommand
 from .diff_report import OverrideAuditDiffReportCommand
 from .diff_single import OverrideAuditDiffSingleCommand
 from .toggle_override import OverrideAuditToggleOverrideCommand
+from .create_override import OverrideAuditCreateOverrideCommand
 from .diff_override import OverrideAuditDiffOverrideCommand
 from .revert_override import OverrideAuditRevertOverrideCommand
 from .diff_externally import OverrideAuditDiffExternallyCommand
@@ -30,6 +32,7 @@ __all__ = [
 
     # Override commands
     "OverrideAuditToggleOverrideCommand",
+    "OverrideAuditCreateOverrideCommand",
     "OverrideAuditDiffOverrideCommand",
     "OverrideAuditDiffExternallyCommand",
     "OverrideAuditRevertOverrideCommand",